[#111] refactor: RecordCreateRequest의 이름을 ExerciseRecordCreateRequest로 변경
