+++ src/main/java/com/example/temp/record/application/ExerciseRecordService.java
@@ -7,7 +7,7 @@ import com.example.temp.member.domain.Member;
 import com.example.temp.member.domain.MemberRepository;
 import com.example.temp.record.domain.ExerciseRecord;
 import com.example.temp.record.domain.ExerciseRecordRepository;
-import com.example.temp.record.dto.request.RecordCreateRequest;
+import com.example.temp.record.dto.request.ExerciseRecordCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -21,7 +21,7 @@ public class ExerciseRecordService {
     private final MemberRepository memberRepository;
 
     @Transactional
-    public long create(UserContext userContext, RecordCreateRequest request) {
+    public long create(UserContext userContext, ExerciseRecordCreateRequest request) {
         Member member = findMember(userContext);
         ExerciseRecord exerciseRecord = request.toEntityWith(member);
         exerciseRecordRepository.save(exerciseRecord);
