+++ src/main/java/com/example/temp/record/presentation/ExerciseRecordController.java
@@ -3,7 +3,7 @@ package com.example.temp.record.presentation;
 import com.example.temp.common.annotation.Login;
 import com.example.temp.common.dto.UserContext;
 import com.example.temp.record.application.ExerciseRecordService;
-import com.example.temp.record.dto.request.RecordCreateRequest;
+import com.example.temp.record.dto.request.ExerciseRecordCreateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -18,7 +18,7 @@ public class ExerciseRecordController {
     private final ExerciseRecordService exerciseRecordService;
 
     @PostMapping
-    public ResponseEntity<Void> create(@Login UserContext userContext, RecordCreateRequest request) {
+    public ResponseEntity<Void> create(@Login UserContext userContext, ExerciseRecordCreateRequest request) {
         long createdId = exerciseRecordService.create(userContext, request);
         return null;
     }
